refactor: Move code from secret-codes to crypto service
