+++ server/src/services/booking.js
@@ -1,19 +1,12 @@
 const mongoose = require('mongoose');
 const BookingModel = mongoose.model('Booking');
 const { fetchPrice } = require('../services/prices');
-const { readKey, signOffer } = require('../services/secret-codes');
 const { sendBookingInfo, sendBookingCanceled } = require('../services/mail');
 const { handleApplicationError } = require('../errors');
-const { generateKeyPair, getKeyPair } = require('../services/crypto');
-const {
-  getCancelBookingTx,
-} = require('../services/web3');
+const { generateKeyPair, getKeyPair, readKey, signOffer } = require('../services/crypto');
+const { getCancelBookingTx } = require('../services/web3');
 const { FROM_EMAIL } = require('../config');
-const {
-  SIGNATURE_TIME_LIMIT,
-  BOOKING_PAYMENT_TYPES,
-  BOOKING_STATUS,
-} = require('../constants');
+const { SIGNATURE_TIME_LIMIT, BOOKING_PAYMENT_TYPES, BOOKING_STATUS } = require('../constants');
 
 async function _generateBooking (data) {
   const index = await BookingModel.nextIndex();
